crawler github: parameterizable PR query to include or not commits data
In case of use an application token to access PR data, if the application
ACLs does not authorize 'contents' read-only then GitHub rejects the graphql
query with commits info requests.

This patch provide a solution to not query commits data. A following patch
will be done to read the application rights to determine if the query need
to be constructed w/o commits request.

Without the commits data we will miss:
- ChangeCommitPushEvent
- ChangeCommitForcePushEvent
- Change populated with each commit details

In the Monocle UI we wont be able to display:
- change_lifecycle histo PR update line
- change_lifecycle iterations by PR
- PR timeline with commits pushed and force pushed events

No patch will be needed at UI level, data will be empty.