github crawler: do not fail on 'The additions count for this commit is unavailable'
This error may happen. I saw it with that PR:
pullrequest.py  --token XXX --org elastic --repository built-docs --id 3

It seems the returned data is complete but w/o the commit (that miss that
'additions' field). So instead of managing that like an error, we
consider it as not fatal and return the data to the caller.

The crawler won't fail on that kind of error.

Fix #222