[web] - changes tables - move away from table and use custom style
The table is not well adapted to custom styling and not to small device
display. This change rework the display of a list of changes.

Also this change removes the ComplexityGraph display for the full list
of open changes. I feel it does not provide enough value and force
user to scroll down the page to see most important data.

Thanks to this it will be easier to add customization in the furture.