[api] Update the Query type to support base expression modifier
This change enables modifying the original Expr data type directly.
This is necessary for the new contributor query where we need to
override the initial dates bound.

This is a good improvement as the new modifier function can be used
in other contexts, for example the ensureMinBound can now work at
the Expr level instead of dealing with the bloodhound specifics.