[api] Update FromJSON HistoSimple instance to work with StrictData
This change uses a Maybe type for histo without sub bucket so that the
decoding works with StrictData.

The instance used a bottom value (error "no subbucket") for the hbSubBuckets
field of SimpleHisto. This no longer works with StrictData because the fields
are strict (where they are always evaluated to normal form), resulting in the
bottom value to throw an exception.

As usual, the solution is to use a correct data type to represent Histo,
that is with a Maybe value to represent the case when there are no sub buckets.