project crawler metedata initialization - use change last updatedAt value at init
In order to ease transition from the legacy Gerrit crawler to the new one, this
patch prevents the full re-indexation of any Gerrit repositories.

The legacy Gerrit crawler do not update the crawler metadata but the new crawler
process is based on it to discover repositories crawling points.

This change inits the crawler metadata (only for project entities) by looking at
the most recent change to grab the updateAt field and set the crawler last_update_at.
If no change is discovered then uses the default crawler date.