[api] run initCrawlerMetadata when no entity to update
This change partially fixes #589.

In case macroscope (crawler) requests the last updated
entity for a given crawler name/type and get Nothing then
we ask the API to run initCrawlerMetadata for that crawler.

However this will work only in case a new crawler is added
and the API has not been restarted.

In case a crawler is amended with changes in the list of
repositories (or organization) to crawl then this solution
is not enough as the API will still return the last updated
entity to refresh and thus will not call initCrawlerMetadata.