[macroscope] ensures stream group key include the Token
This change updates the getClientGraphQL to include the token
value inside the stream group name.

Indeed mulitple tokens could be used with a single url and to
prepare for handling the query quota (handled by Token) we need to
separate group by url/token as streams inside a group are
processed sequentialy.