[macroscope] Lentille.GitHub.PullRequests - use search query type
The respository query does not ease to specify on particular
Pull Request making the request not practical to be used in
debugging case where we need to query a specific PR.

The PullRequests and PullRequest attributes in repository query
makes code factorization un-practical because Morpheus generates
two similar but incompatible data types.

This is a change compared to the current way of querying GitHub
PRs (in legacy crawler) but that that might be worth to try this
new way.