[macroscope] github crawler - PR - retry on BadGateway "Something went wrong ..."
This commits ensures that Macroscope will retry GitHub GraphQL requests
that fails during decoding due to a server side timeout.

This timeout occurs when too much data are asked to the GitHub GraphQL
API so that the query take longer than (authorized?).

When a timeout occurs query response is a 502 BadGateway with a payload
message such as:

"Something went wrong while executing your query. This may be the result
of a timeout ..."

To do so, this commit implements:
- a generic retry function for an IO action that can be retry based on a
  provided function that inspect IO action result.
- a new function (doRequest) that wraps fetchWithLog with genericRetry
  based on a check function provided by the GitHub.PullRequests module.
- doRequest is able to decrease an optional depth value passed to
  the graphQL arguments builder.