[webui] re-add group members listing and remove store usage
This patch add group member listing inside the scoped group view
and make member name clickable.

Also and mainly this patch removes the store usage to store the
scope type and entity to only rely on component props and
route props. This makes the management less error prone and make
it more friendly for the browser history.