[api] fix missusage of author filter for TOP_COMMENTED/TOP_REVIEWED queries
Those queries list the most commented or reviewed change authors, but
when an author filter is specified we expect to get filtered on the
comment or review events created by the filtered author.

In other words, by filtering on an author we want to see on which change
authors a reviewer or commenter have been the most active.