[auth] resolve Monocle UID based on Provider JWT claim
This patch resolves the Monocle UID from the Provider JWT
user claim.

Before creating the local JWT we look for matching Monocle
UID(s) (ident) over all indexes.

The JWT contains two values a default Monocle UID which is the
content of the user claim and a mapping of (index, ident).
When the mapping is empty or the lookup for a given index fails
then the default Monocle UID is used.