[auth] Do not reset local JWK material at restart
This change provides a solution to build the JWK Material
from a bytes string passed via an environment variable
(MONOCLE_JWK_GEN_KEY). The minimal size of the this bytes string
must be 64 bytes.

If not provided or less than 64 Bytes then Monocle generates
a random bytes string however when the Monocle API is restarted
the JWK material change then already delivered tokens won't
be validated by the API.