Introduce Monocle.Entity module
This change refactors the code so that crawler metadata entity
are no longer converted to text outside of the json encoding.
This improves the crawler metadata usage by removing the invariant
that an entity can be coverted to a (entity type, entity value) text pair.

For example, a project entity is presently encoded as:
 {"crawler_type": "project", "crawler_type_value": "project-name"}

The goal is to enable richer entity definition, for example:
 {"entity_type": "issue", "entity_project", "project-name", "entity_issue_number": "42"}