crawler - Github PR - relax ratelimit in case deactivated in GHE
According to #958, the rate limit system can be deactivated in GHE.

This change updates the Github PR crawler to handle a ratelimit
returned as null by the API. In this case Macroscope's GraphQL stream
handler does not check for the internal RateLimit object because
stored as Nothing then the waiting delay to honor a ratelimit is
skipped.

Related to issue: #958

This commit does not handle the case for other Github crawlers.
If this is accepted some follow-up patches will be needed to
handle this is other Github crawlers.