Web - enable HTMX
The idea is to get a first simple hook component that make it
get a HTMX resource on the server based on an event.

Here is an usage example:

```
<HTMXGetHook url={"/htmx/test"} trigger={"click"}>
    <button> {"Hello HTMX"->str} </button>
</HTMXGetHook>
```

In the monocle use case, the trigger event might be `load` with
the component's children to React.Null in order to make the browser
fetch an htmx element from the server just after the component loaded.