Remove local-storage usage for JWT
This change makes use of an additional (non httpOnly) session cookie
to store non sensitive data related to the autenticated user.
The data inside the "Monocle" cookie are serialized into an
already expired JWT.

The Web app read the Monocle Cookie to extract authenticated
user information. This was previously done by reading the JWT
from the local storage.

Furthermore, the XSRF token is deactivated from servant-auth-server
as it seems that the high requests rate that the Monocle UI
perform on the server (like on the User page) creates a race
condition leading that for some requests the X-XSRF-TOKEN header
and XSRF-Cookie values differ then the API return 403. Unfortunatly
I was unable to find the proper fix. Some aspect mitigate the
removal like the use of sameSite: Lax for the JWT-Cookie session
cookie and and the use of JSX, furthermore Monocle does not store
sensitive information.