Add Lentille.GitHub.UserPullRequests module
The module provides a streamUserPullRequests function to stream PRs
from a given github username with a date limit.

This change brings a bit of duplication due to newtypes generated
by Morpheus client (query name is different) but we tried to
mitigate by moving some generic/duplicated code in a Utils module.

Also a CLI command is added to run user PR stream.