Update peer strength metric computation to enable filtering by change author
Fixes: #1037

This change updates the computation where now we:

Find (and filter on) changes authors and for each of them find
all event authors that did a comment/review on the change author.

The previous computation was that:

Find (and filter on) comments/reviews events authors and for each of
them find all changes authors on which the event's author did a
comment/review.

This change also:
- Updates the web UI for the peer strength to set the author colunm on
  the left.
- Keeps the previous computation code but unused for now. If the previous
  computation is relevant we could add a switch in the UI.