Ensure Github PR fetching repo not found non fatal for stream
When Monocle discovers organization's repositories it might happen
that later repositories are deleted then the Changes crawler may
encounter GraphQL queries issues. Such case in now handled
and the stream silently continue on that case.

This effort could be subject to followup. But this first change
might solve the issue #1112